Feature/#23

+++ src/main/java/com/skyhorsemanpower/payment/payment/dto/PaymentListResponseDto.java
@@ -2,6 +2,7 @@ package com.skyhorsemanpower.payment.payment.dto;
 
 import com.skyhorsemanpower.payment.common.PaymentStatus;
 import com.skyhorsemanpower.payment.payment.vo.PaymentListResponseVo;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,19 +17,19 @@ import lombok.Setter;
 @Builder
 public class PaymentListResponseDto {
 
-	private String paymentUuid;
-	private String auctionUuid;
-	private double price;
-	private PaymentStatus paymentStatus;
-	private LocalDateTime paymentAt;
+    private String paymentUuid;
+    private String auctionUuid;
+    private BigDecimal price;
+    private PaymentStatus paymentStatus;
+    private LocalDateTime paymentAt;
 
-	public static PaymentListResponseVo dtoToVo(PaymentListResponseDto paymentListResponseDto) {
-		return new PaymentListResponseVo(
-			paymentListResponseDto.getPaymentUuid(),
-			paymentListResponseDto.getAuctionUuid(),
-			paymentListResponseDto.getPrice(),
-			paymentListResponseDto.getPaymentStatus(),
-			paymentListResponseDto.getPaymentAt());
-	}
+    public static PaymentListResponseVo dtoToVo(PaymentListResponseDto paymentListResponseDto) {
+        return new PaymentListResponseVo(
+            paymentListResponseDto.getPaymentUuid(),
+            paymentListResponseDto.getAuctionUuid(),
+            paymentListResponseDto.getPrice(),
+            paymentListResponseDto.getPaymentStatus(),
+            paymentListResponseDto.getPaymentAt());
+    }
 
 }
