+++ src/main/java/com/skyhorsemanpower/payment/payment/vo/PaymentAddRequestVo.java
@@ -1,12 +1,13 @@
 package com.skyhorsemanpower.payment.payment.vo;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 
 @Getter
 public class PaymentAddRequestVo {
 
-	private String auctionUuid;
-	private String paymentMethod;
-	private double price;
-	private String paymentNumber;
+    private String auctionUuid;
+    private String paymentMethod;
+    private BigDecimal price;
+    private String paymentNumber;
 }
